Added transaction management to be able to delete big tree (file folder structure) from trash can

+++ alfrescotrashcancleaner2/src/main/java/org/alfresco/trashcan/cleaner/TrashcanCleaner.java
@@ -17,6 +17,7 @@ import org.alfresco.repo.lock.LockAcquisitionException;
 import org.alfresco.repo.model.filefolder.GetChildrenCannedQueryFactory;
 import org.alfresco.repo.node.getchildren.GetChildrenCannedQuery;
 import org.alfresco.repo.transaction.RetryingTransactionHelper.RetryingTransactionCallback;
+import org.alfresco.service.cmr.repository.ChildAssociationRef;
 import org.alfresco.service.cmr.repository.NodeRef;
 import org.alfresco.service.cmr.repository.NodeService;
 import org.alfresco.service.cmr.repository.StoreRef;
@@ -168,12 +169,56 @@ public class TrashcanCleaner
                                 logger.debug("Delete NodeRef :" + nodeRef + " $ Name :" + name + " $ archivedDate :"
                                         + archivedDate);
                             }
-                            nodeService.deleteNode(nodeRef);
-
+                            //Maybe there is more than one element to deleted there
+                            //It might be a big three and therefore causing  requiring big or even too big transaction
+                            //nodeService.deleteNode(nodeRef);
+                            
+                            while(deleteRecursive(nodeRef, 500) != 0);
                         }
 
                         return true;
                     }
+                    
+                    int deleteRecursive(NodeRef nodeRef, int maxNumOfNodesInTransaction)
+                    {
+                        final NodeRef fNodeREf = nodeRef;
+                        final int fMaxNumOfNodesInTransaction = maxNumOfNodesInTransaction;
+                        final RetryingTransactionCallback<Integer> recursiveWork = new RetryingTransactionCallback<Integer>()
+                                {
+                                    public Integer execute() throws Exception
+                                    {
+                                        return this.deleteRecursiveInternal(fNodeREf,500);
+                                    }
+                                    
+                                    public int deleteRecursiveInternal(NodeRef nodeRef, int fMaxNumOfNodesInTransaction)
+                                    {
+                                        if (!nodeService.exists(nodeRef))
+                                            return 0; //finished
+                                        List<ChildAssociationRef> allChildren = nodeService.getChildAssocs(nodeRef);
+                                        
+                                        //if no child then delete the node, this is a leaf
+                                        if ( allChildren == null || allChildren.size() == 0)
+                                        {
+                                           nodeService.deleteNode(nodeRef);
+                                           return 1;
+                                        }
+                                        
+                                        //recursive an iterate and count
+                                        int sum = 0;
+                                        for(ChildAssociationRef assoc: allChildren)
+                                        {
+                                            sum += deleteRecursiveInternal(assoc.getChildRef(), fMaxNumOfNodesInTransaction);
+                                            if (sum >= fMaxNumOfNodesInTransaction)
+                                                break;
+                                        }
+                                        return sum;
+                                    }
+                                
+                                };
+                                return transactionService.getRetryingTransactionHelper().doInTransaction(recursiveWork, false, true);
+                    }
+                    
+
                 };
 
             final AtomicBoolean keepGoing = new AtomicBoolean(true);
@@ -234,4 +279,6 @@ public class TrashcanCleaner
 
         }
     }
+    
+
 }
